Feature: Added new json responses
