+++ canary/go.mod
@@ -1,0 +1,3 @@
+module github.com/ajmclaug1/simple-go-web-app/canary
+
+go 1.20
