+++ canary/static/index.html
@@ -1,0 +1,63 @@
+<!DOCTYPE HTML>
+<html>
+   <head>
+      <title>Canary demo</title>
+      <meta http-equiv="content-type" content="text/html; charset=utf-8" />
+      <link rel="shortcut icon" href="favicon.png" type="image/png" />
+      <link rel="stylesheet" type="text/css" href="style.css"/>
+      <title>Rounded Corner Boxes</title>
+      <script src="https://unpkg.com/htmx.org@1.4.1"></script>
+   </head>
+   <body>
+       <h1> Request dashboard</h1>
+      <div class="box" hx-get="callme" hx-trigger="every 3s">
+         <h2>Pod request</h2>
+      </div>
+      <div class="box" hx-get="callme" hx-trigger="every 3s">
+         <h2>Pod request</h2>
+      </div>
+      <div class="box" hx-get="callme" hx-trigger="every 3s">
+         <h2>Pod request</h2>
+      </div>
+      <div class="box" hx-get="callme" hx-trigger="every 3s">
+         <h2>Pod request</h2>
+      </div>
+      <div class="box" hx-get="callme" hx-trigger="every 3s">
+         <h2>Pod request</h2>
+      </div>
+      <div class="box" hx-get="callme" hx-trigger="every 3s">
+         <h2>Pod request</h2>
+      </div>
+      <div class="box" hx-get="callme" hx-trigger="every 3s">
+         <h2>Pod request</h2>
+      </div>
+      <div class="box" hx-get="callme" hx-trigger="every 3s">
+         <h2>Pod request</h2>
+      </div>
+      <div class="box" hx-get="callme" hx-trigger="every 3s">
+         <h2>Pod request</h2>
+      </div>
+      <div class="box" hx-get="callme" hx-trigger="every 3s">
+         <h2>Pod request</h2>
+      </div>
+      <div class="box" hx-get="callme" hx-trigger="every 3s">
+         <h2>Pod request</h2>
+      </div>
+      <div class="box" hx-get="callme" hx-trigger="every 3s">
+         <h2>Pod request</h2>
+      </div>
+      <div class="box" hx-get="callme" hx-trigger="every 3s">
+         <h2>Pod request</h2>
+      </div>
+      <div class="box" hx-get="callme" hx-trigger="every 3s">
+         <h2>Pod request</h2>
+      </div>
+      <div class="box" hx-get="callme" hx-trigger="every 3s">
+         <h2>Pod request</h2>
+      </div>
+      <div class="box" hx-get="callme" hx-trigger="every 3s">
+         <h2>Pod request</h2>
+      </div>
+   </body>
+</html>
+
