--- go-web-service/main.go
@@ -1,50 +1,0 @@
-package main
-
-import (
-	"flag"
-	"fmt"
-	"log"
-	"net/http"
-	"os"
-	"time"
-)
-
-const version = "1.0.0"
-
-type config struct {
-	port int
-	env  string
-}
-
-type application struct {
-	config config
-	logger *log.Logger
-}
-
-func main() {
-	var cfg config
-
-	flag.IntVar(&cfg.port, "port", 4000, "API server port")
-	flag.StringVar(&cfg.env, "env", "dev", "Environment (dev|stage|prod)")
-	flag.Parse()
-
-	logger := log.New(os.Stdout, "", log.Ldate|log.Ltime)
-	app := &application{
-		config: cfg,
-		logger: logger,
-	}
-
-	addr := fmt.Sprintf(":%d", cfg.port)
-
-	srv := &http.Server{
-		Addr:         addr,
-		Handler:      app.route(),
-		IdleTimeout:  time.Minute,
-		ReadTimeout:  10 * time.Second,
-		WriteTimeout: 30 * time.Second,
-	}
-
-	logger.Printf("starting %s server on %s", cfg.env, addr)
-	err := srv.ListenAndServe()
-	logger.Fatal(err)
-}
