+++ simple-go-server/internal/data/books.go
@@ -1,0 +1,16 @@
+package data
+
+import(
+  "time"
+)
+type Book struct {
+
+  ID int64
+  CreatedAt time.Time
+  Title string
+  Published int64
+  Pages int
+  Genres []string
+  Rating float32
+  Version int32
+}
