+++ simple-go-server/routes.go
@@ -1,0 +1,11 @@
+package main
+
+import "net/http"
+
+func (app *application) route() *http.ServeMux {
+	mux := http.NewServeMux()
+	mux.HandleFunc("/v1/healthcheck", app.healthcheck)
+	mux.HandleFunc("/v1/books", app.getCreateBooksHandler)
+	mux.HandleFunc("/v1/books/", app.getUpdateDeleteBooksHandler)
+	return mux
+}
